+++ inbm/Changelog.md
@@ -4,8 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## NEXT - YYYY-MM-DD
+## Next Release
 ### Changed
+ - (NEXMANAGE-513) Record the granular log in download-only mode
  - (NEX-15262) Returns failure reason that matches the format required by MM
+   
+### Added
+ - (NEXMANAGE-826) Update granular log to support architecture independent package
 
 ## 4.2.7 - 2024-11-8
 ### Changed